bigggg nataru update

+++ app/Http/Controllers/KelasController.php
@@ -6,6 +6,7 @@ use App\Models\Kelas;
 use App\Models\Pegawai;
 use App\Models\Siswa;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class KelasController extends Controller
 {
@@ -14,6 +15,10 @@ class KelasController extends Controller
      */
     public function index()
     {
+        if (!Gate::any(['staf-tata-usaha', 'guru'])) {
+            abort(404);
+        }
+
         $kelas = Kelas::with('pegawai')->filter(request()->except('kelas_filter'))->paginate(20)->withQueryString();
         $siswa = Siswa::all();
 
@@ -29,6 +34,10 @@ class KelasController extends Controller
      */
     public function create()
     {
+        if (!Gate::allows('staf-tata-usaha')) {
+            abort(404);
+        }
+
         $guru = Pegawai::where('posisi', 'Guru')->latest()->get();
 
         return view('pages.master.kelas.create', [
@@ -42,6 +51,10 @@ class KelasController extends Controller
      */
     public function store(Request $request)
     {
+        if (!Gate::allows('staf-tata-usaha')) {
+            abort(404);
+        }
+
         $kelas_validation_rules = [
             'nama_kelas' => 'required|string|min:2|max:5|unique:kelas,nama_kelas|regex:/^(?=.*[A-Za-z])(?=.*[0-9])[A-Za-z0-9]{2,5}$/',
             'id_pegawai' => 'nullable|integer|exists:pegawai,id_pegawai|unique:kelas,id_pegawai',
@@ -59,6 +72,10 @@ class KelasController extends Controller
      */
     public function show(Kelas $kelas)
     {
+        if (!Gate::any(['staf-tata-usaha', 'guru'])) {
+            abort(404);
+        }
+
         return view('pages.master.kelas.show', [
             'judul' => 'Kelas',
             'kelas' => $kelas
@@ -70,6 +87,10 @@ class KelasController extends Controller
      */
     public function edit(Kelas $kelas)
     {
+        if (!Gate::allows('staf-tata-usaha')) {
+            abort(404);
+        }
+
         $guru = Pegawai::where('posisi', 'Guru')->latest()->get();
 
         return view('pages.master.kelas.edit', [
@@ -84,6 +105,10 @@ class KelasController extends Controller
      */
     public function update(Request $request, Kelas $kelas)
     {
+        if (!Gate::allows('staf-tata-usaha')) {
+            abort(404);
+        }
+
         $kelas_validation_rules_update = [
             'nama_kelas' => "required|string|min:2|max:5|unique:kelas,nama_kelas,{$kelas->id_kelas},id_kelas|regex:/^(?=.*[A-Za-z])(?=.*[0-9])[A-Za-z0-9]{2,5}$/",
             'id_pegawai' => "nullable|integer|exists:pegawai,id_pegawai|unique:kelas,id_pegawai,{$kelas->id_kelas},id_kelas"
@@ -101,6 +126,10 @@ class KelasController extends Controller
      */
     public function destroy(Kelas $kelas)
     {
+        if (!Gate::allows('staf-tata-usaha')) {
+            abort(404);
+        }
+        
         $kelas->delete();
 
         return redirect()->route('kelas.index')->with('success', 'Kelas berhasil dihapus.');
