+++ app/Http/Controllers/PrestasiController.php
@@ -6,6 +6,8 @@ use App\Models\Prestasi;
 use App\Models\Siswa;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Support\Facades\Auth;
 
 class PrestasiController extends Controller
 {
@@ -27,9 +29,15 @@ class PrestasiController extends Controller
      */
     public function index()
     {
-        $prestasi = Prestasi::with('siswa')->filter(request()->all())->paginate(20)->withQueryString();
+        if (Gate::any(['staf-tata-usaha', 'guru']))
+            $prestasi = Prestasi::with('siswa')->filter(request()->all())->paginate(20)->withQueryString();
+        else if (Gate::allows('siswa')) {
+            $prestasi = Prestasi::with('siswa')->where('id_siswa', Auth::user()->siswa->id_siswa)->filter(request()->all())->paginate(20)->withQueryString();
+        } else {
+            abort(404);
+        }
 
-        $siswa = Siswa::latest()->get();
+        $siswa = Siswa::get();
 
         return view('pages.akademik.prestasi.index', [
             'judul' => 'Prestasi',
@@ -43,6 +51,10 @@ class PrestasiController extends Controller
      */
     public function create()
     {
+        if (!Gate::allows('staf-tata-usaha')) {
+            abort(404);
+        }
+
         $siswa = Siswa::latest()->get();
 
         return view('pages.akademik.prestasi.create', [
@@ -56,6 +68,10 @@ class PrestasiController extends Controller
      */
     public function store(Request $request)
     {
+        if (!Gate::allows('staf-tata-usaha')) {
+            abort(404);
+        }
+
         $validated_prestasi = $request->validate($this->prestasi_validation_rules);
 
         if ($request->hasFile('dokumentasi')) {
@@ -86,6 +102,10 @@ class PrestasiController extends Controller
      */
     public function edit(Prestasi $prestasi)
     {
+        if (!Gate::allows('staf-tata-usaha')) {
+            abort(404);
+        }
+
         return view('pages.akademik.prestasi.edit', [
             'judul' => 'Prestasi',
             'prestasi' => $prestasi
@@ -97,6 +117,10 @@ class PrestasiController extends Controller
      */
     public function update(Request $request, Prestasi $prestasi)
     {
+        if (!Gate::allows('staf-tata-usaha')) {
+            abort(404);
+        }
+
         $prestasi_validation_rultes_update = $this->prestasi_validation_rules;
         $prestasi_validation_rultes_update['image_delete'] = 'required|integer';
 
@@ -130,6 +154,10 @@ class PrestasiController extends Controller
      */
     public function destroy(Prestasi $prestasi)
     {
+        if (!Gate::allows('staf-tata-usaha')) {
+            abort(404);
+        }
+        
         $prestasi->delete();
 
         if (!empty($prestasi->dokumentasi)) {
